Add 'roundtripping' to rle_test.
This simply takes the output of a test,
re-(de)compresses it, and compares the result
to the original input, which should usually match.

For the few cases where this operation is expected
to fail, like for instance any input containing
NOPs or REP 0, the flag '-' can be added to the
test action to disable roundtripping.